gdata: Set the correct nco:creator for PicasaWeb
We use the email address to uniquely identify each nco:Contact, and
the gdata_entry_get_authors API does not work correctly for PicasaWeb.
The GDataAuthor obtained from gdata_entry_get_authors for a
GDataPicasaWebAlbum has a NULL email address. This confuses the code
to pick a random nco:Contact as the creator of the album.

Using the data source's URN as the fake email address of an nco:Contact
is also wrong because it will be the same for everything across a
single GOA account.

Instead, we can use the MD5 checksum of the fullname to uniquely
identity each nco:Contact.

https://bugzilla.gnome.org/show_bug.cgi?id=740423