gdata: Page through the GDataDocumentsFeed
There can be a very large number of documents in a Google Drive.
Getting all of them at once is not a good idea because there can be an
upper bound on how long the response from the server can be.

Only the failure to get the first page is treated as fatal. Failure to
get any subsequent page terminates the operation but we don't report
the failure to the caller.

https://bugzilla.gnome.org/show_bug.cgi?id=770656