miner: Get the cancellable from the GTask, not an instance variable
Just as each invocation of gom_miner_refresh_db_async should have a
separate GTask, it should also use the GCancellable specifically passed
by the caller. Having an instance variable in the GomMiner class
violates this principle.

There is also no need for a separate GCancellable in
GomAccountMinerJob. We can use the one passed in by the caller.

https://bugzilla.gnome.org/show_bug.cgi?id=770889