zpj: Use an HTTP URL as nie:url
The nao:identifier value is not a valid nie:url. The identifier has the
format "windows-live:skydrive:<id>", and is not a GVfs supported URI.
This is a problem because the nie:url is used to locate the path to
the entry's thumbnail in the desktop's thumbnail cache. Running
g_file_query_info (file, G_FILE_ATTRIBUTE_THUMBNAIL_PATH, ...) on such
a URI will fail. Therefore we use a fake OneDrive-namespaced URI that
GVfs understands.

Force a full reindex of OneDrive resources for existing users.